pyproject.toml
[project]
name = "nexus"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13,<3.14"
dependencies = [
    "aiokafka>=0.13.0",
    "fastapi>=0.123.10",
    "google-adk[extensions]>=1.22.1",
    "google-genai>=1.59.0",
    "google-generativeai>=0.8.6",
    "litellm>=1.75.5",
    "mistralai>=1.9.11",
    "pillow>=12.1.0",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "qdrant-client>=1.16.2",
    "rank-bm25>=0.2.2",
    "sentence-transformers>=5.2.0",
    "spacy>=3.8.11",
    "torch>=2.9.1",
    "transformers>=4.57.6",
    "uvicorn>=0.40.0",
]
